pyproject.toml
[project]
name = "fastapi-model-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "accelerate==0.33.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.0",
    "aiosignal==1.4.0",
    "airportsdata==20250909",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "astor==0.8.1",
    "async-timeout==5.0.1",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.2",
    "blake3==1.0.7",
    "build==1.3.0",
    "cachetools==6.2.0",
    "certifi==2025.10.5",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "cmake==4.1.0",
    "compressed-tensors==0.9.3",
    "cupy-cuda12x==13.6.0",
    "deprecated==1.2.18",
    "depyf==0.18.0",
    "diffusers==0.32.2",
    "dill==0.4.0",
    "diskcache==5.6.3",
    "distlib==0.4.0",
    "distro==1.9.0",
    "distvae==0.0.0b5",
    "dnspython==2.8.0",
    "einops==0.8.1",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.118.3",
    "fastapi-cli==0.0.13",
    "fastapi-cloud-cli==0.3.1",
    "fastrlock==0.8.3",
    "filelock==3.20.0",
    "frozenlist==1.8.0",
    "fsspec==2025.9.0",
    "ftfy==6.3.1",
    "gguf==0.17.1",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.75.1",
    "h11==0.16.0",
    "hf-xet==1.1.10",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==0.35.3",
    "identify==2.6.15",
    "idna==3.10",
    "importlib-metadata==8.0.0",
    "iniconfig==2.1.0",
    "interegular==0.3.3",
    "jinja2==3.1.6",
    "jiter==0.11.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "lark==1.2.2",
    "lit==18.1.8",
    "llguidance==0.7.30",
    "llvmlite==0.44.0",
    "lm-format-enforcer==0.10.12",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mistral-common==1.8.5",
    "modelscope==1.30.0",
    "mpmath==1.3.0",
    "msgpack==1.1.2",
    "msgspec==0.19.0",
    "multidict==6.7.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "ninja==1.13.0",
    "nodeenv==1.9.1",
    "numba==0.61.2",
    "numpy==1.26.4",
    "nvidia-cublas-cu11==11.11.3.6",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu11==11.8.87",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu11==11.8.89",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu11==11.8.89",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu11==9.1.0.70",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu11==10.9.0.58",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu11==10.3.0.86",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu11==11.4.1.48",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu11==11.7.5.86",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu11==2.20.5",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu11==11.8.86",
    "nvidia-nvtx-cu12==12.4.127",
    "openai==2.3.0",
    "opentelemetry-api==1.26.0",
    "opentelemetry-exporter-otlp==1.26.0",
    "opentelemetry-exporter-otlp-proto-common==1.26.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.26.0",
    "opentelemetry-exporter-otlp-proto-http==1.26.0",
    "opentelemetry-proto==1.26.0",
    "opentelemetry-sdk==1.26.0",
    "opentelemetry-semantic-conventions==0.47b0",
    "opentelemetry-semantic-conventions-ai==0.4.13",
    "outlines==0.1.11",
    "outlines-core==0.1.26",
    "packaging==25.0",
    "partial-json-parser==0.2.1.1.post6",
    "pathspec==0.12.1",
    "pillow==11.3.0",
    "pip==25.2",
    "platformdirs==4.5.0",
    "pluggy==1.6.0",
    "pre-commit==4.3.0",
    "prometheus-client==0.23.1",
    "prometheus-fastapi-instrumentator==7.1.0",
    "propcache==0.4.1",
    "protobuf==4.25.8",
    "psutil==7.1.0",
    "py-cpuinfo==9.0.0",
    "pycountry==24.6.1",
    "pydantic==2.12.0",
    "pydantic-core==2.41.1",
    "pydantic-extra-types==2.10.6",
    "pygments==2.19.2",
    "pyproject-hooks==1.2.0",
    "pytest==8.4.2",
    "python-dotenv==1.1.1",
    "python-json-logger==4.0.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.3",
    "pyzmq==26.2.0",
    "ray==2.50.0",
    "referencing==0.36.2",
    "regex==2025.9.18",
    "requests==2.32.5",
    "rich==14.2.0",
    "rich-toolkit==0.15.1",
    "rignore==0.7.0",
    "rpds-py==0.27.1",
    "ruff==0.14.0",
    "safetensors==0.6.2",
    "scikit-build-core==0.11.6",
    "scipy==1.15.3",
    "semantic-version==2.10.0",
    "sentencepiece==0.1.99",
    "sentry-sdk==2.41.0",
    "setuptools==80.9.0",
    "setuptools-rust==1.12.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "starlette==0.48.0",
    "sympy==1.13.1",
    "tiktoken",
    "tokenizers==0.21.4",
    "tomli==2.3.0",
    "torch==2.4.1+cu118",
    "torchaudio==2.4.1+cu118",
    "torchvision==0.19.1+cu118",
    "tqdm==4.67.1",
    "transformers==4.49.0",
    "triton==3.0.0",
    "typer==0.19.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uvicorn[standard]==0.37.0",
    "uvloop==0.21.0",
    "virtualenv==20.35.1",
    "watchfiles==1.1.0",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
    "wheel==0.45.1",
    "wrapt==1.17.3",
    "xfuser==0.4.4",
    "xgrammar==0.1.18",
    "yarl==1.22.0",
    "yunchang==0.6.3.post1",
    "zipp==3.23.0",
]
[tool.uv]
override-dependencies = [
    "opencv-python-headless<4.12"
]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv.sources]
tiktoken = { path = "tiktoken-0.11.0-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl" }

